support extraEnvVars, extraEnvVarsCM, extraEnvVarsSecret (#10)
This change adds support for an `extraEnvVars` value to set env vars to be
used in the rendered deployment and its pods. It also adds
`extraEnvVarCM` for getting env vars from a configmap and
`extraEnvVarsSecret` for getting them from a secret dynamically.

As it says in the included values.yaml comment, it takes a list of dicts
with name and value:

```
extraEnvVars:
  - name: LOG_LEVEL
    value: "debug"
  - name: CUSTOM_VAR
    value: "custom-value"
```

And if you want to specify a configmap or secret, it can be done like
this:

```
extraEnvVarsCM: "mymap"
extraEnvVarsSecret: "mysecret"
```